bitkeeper revision 1.930 (40bf0f4cD6X2TCNUTUzQPA_qDsagoQ)
Reduce domid to 32 bits only -- more than enough.

+++ tools/examples/xc_dom_control.py
@@ -50,7 +50,7 @@ dom = None
 
 
 if len( sys.argv ) > 2 and re.match('\d+$', sys.argv[2]):
-    dom = long(sys.argv[2])
+    dom = int(sys.argv[2])
 
 if cmd == 'stop':
     rc = xc.domain_stop( dom=dom )
