+++ tools/xc/lib/xc_linux_build.c
@@ -21,7 +21,7 @@ static int readelfimage_base_and_size(char *elfbase,
 static int loadelfimage(char *elfbase, int pmh, unsigned long *parray,
                         unsigned long vstart);
 
-static long get_tot_pages(int xc_handle, u64 domid)
+static long get_tot_pages(int xc_handle, u32 domid)
 {
     dom0_op_t op;
     op.cmd = DOM0_GETDOMAININFO;
@@ -32,7 +32,7 @@ static long get_tot_pages(int xc_handle, u64 domid)
 }
 
 static int get_pfn_list(int xc_handle,
-                        u64 domid, 
+                        u32 domid, 
                         unsigned long *pfn_buf, 
                         unsigned long max_pfns)
 {
@@ -66,12 +66,12 @@ static int copy_to_domain_page(int pm_handle,
 }
 
 static int setup_guestos(int xc_handle,
-                         u64 dom,
+                         u32 dom,
                          char *image, unsigned long image_size,
                          gzFile initrd_gfd, unsigned long initrd_len,
                          unsigned long nr_pages,
                          unsigned long *pvsi, unsigned long *pvke,
-			 full_execution_context_t *ctxt,
+                         full_execution_context_t *ctxt,
                          const char *cmdline,
                          unsigned long shared_info_frame,
                          unsigned int control_evtchn)
@@ -377,7 +377,7 @@ static char *read_kernel_image(const char *filename, unsigned long *size)
 }
 
 int xc_linux_build(int xc_handle,
-                   u64 domid,
+                   u32 domid,
                    const char *image_name,
                    const char *ramdisk_name,
                    const char *cmdline,
@@ -429,7 +429,7 @@ int xc_linux_build(int xc_handle,
     op.u.getdomaininfo.domain = (domid_t)domid;
     op.u.getdomaininfo.ctxt = ctxt;
     if ( (do_dom0_op(xc_handle, &op) < 0) || 
-         ((u64)op.u.getdomaininfo.domain != domid) )
+         ((u32)op.u.getdomaininfo.domain != domid) )
     {
         PERROR("Could not get info on domain");
         goto error_out;
